% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMICToGrowth.R
\name{addMICToGrowth}
\alias{addMICToGrowth}
\title{Add MIC values to Growth results}
\usage{
addMICToGrowth(namedGrowthList, growthThreshold = 0.1,
  growthControlCol = 11, blankCol = 12, maxConc, dilutionStep)
}
\arguments{
\item{namedGrowthList}{named list element from the output of convertODDataToGrowth()}
}
\description{
This function allows you to convert a table containing the OD values from one
named list element to a table containing %growth. The output is a named list
of length 1
}
\keyword{OD,}
\keyword{Tecan}
\keyword{growth}
\keyword{output,}
\keyword{percentage}
