% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGrowthFraction.R
\name{calculateGrowthFraction}
\alias{calculateGrowthFraction}
\title{Convert optical denisty into growth fraction}
\usage{
calculateGrowthFraction(wellOpticalDensity, growthControlOpticalDensity,
  blankOpticalDensity, siginifDigits = 3)
}
\arguments{
\item{wellOpticalDensity}{optical density measured in the well of interest}

\item{growthControlOpticalDensity}{optical density measured in the growth
control well}

\item{blankOpticalDensity}{optical density measured in the blank well}

\item{siginifDigits}{number of significants digits for the growth fraction. Defaults to 3}
}
\description{
This function calculates the % of growth observed in a well using the optical
densities measured in that well and in a growth control and a blank well
}
\examples{
calculateGrowthFraction(0.2,0.5,0.03)
0.362
}
\keyword{Optical}
\keyword{density,}
\keyword{fraction}
\keyword{growth}
