% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concSerie.R
\name{concSerie}
\alias{concSerie}
\title{Calculate geometric serie of concentrations}
\usage{
concSerie(maxConc, dilutionStep = 2, numberOfConc = 10)
}
\arguments{
\item{maxConc}{drug concentration in the most concentrated well}

\item{dilutionStep}{dilution step defaults to 2}

\item{numberOfConc}{number of wells, defaults to 10}
}
\description{
This function allows you to calculate the drug concentration present in each
well of a line by inputing the concentration of the most concentrated well
the number of wells in the line and the dilution step
}
\keyword{OD,}
\keyword{Tecan}
\keyword{growth}
\keyword{output,}
\keyword{percentage}
