% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGrowthToBinary.R
\name{convertGrowthToBinary}
\alias{convertGrowthToBinary}
\title{Convert % growth tables to growth/no growth tables}
\usage{
convertGrowthToBinary(namedGrowthList, growthThreshold = 0.1)
}
\arguments{
\item{namedGrowthList}{named list element from the output of convertODDataToGrowth()}

\item{growthThreshold}{threshold above which we consider that growth is achived. Defaults to 0.1 (10%)}
}
\description{
This function allows you to convert a list of tables containing %growth from one
named list element to a list of tables containing growth/no growth info. The output is a named list
}
\keyword{OD,}
\keyword{Tecan}
\keyword{growth}
\keyword{output,}
\keyword{percentage}
