% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertODDataToGrowth.R
\name{convertODDataToGrowth}
\alias{convertODDataToGrowth}
\title{Convert OD tables to % growth tables}
\usage{
convertODDataToGrowth(namedODList, growthControlCol = 11,
  blankCol = 12, siginifDigits = 3)
}
\arguments{
\item{growthControlCol}{number of the column containing the growth control. Default is 11}

\item{blankCol}{number of the column containing the broth control. Default is 12}

\item{siginifDigits}{number of significants digits for the growth fraction. Defaults to 3}

\item{namedODdata}{named list element from the output of extractODTecanSheet()}
}
\description{
This function allows you to convert a list of tables containing the OD values from one
named list element to a table containing %growth. The output is a named list
}
\keyword{OD,}
\keyword{Tecan}
\keyword{growth}
\keyword{output,}
\keyword{percentage}
