% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMICTables.R
\name{createMICTables}
\alias{createMICTables}
\title{Convert plate plans and MIC columns index to MIC values}
\usage{
createMICTables(MICColumns, platePlans, plateOptions)
}
\arguments{
\item{MICColumns}{named list of data frames containing the columns index of MIC}

\item{platePlans}{named list of data frames containing plate plans}

\item{plateOptions}{named list of data frames containing plate options}
}
\description{
This function allows you to convert a list of tables containing the index of columns conatining
the MIC concentration and a list of plate plans to a list of tables containing MIC values
The output is a named list
}
\keyword{OD,}
\keyword{Tecan}
\keyword{growth}
\keyword{output,}
\keyword{percentage}
