% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlatePlan.R
\name{createPlatePlan}
\alias{createPlatePlan}
\title{Updates an already existing plate plan}
\usage{
createPlatePlan(inputTable, colGrowthCtl = 11, colBrothCtl = 12,
  nlines = 8)
}
\arguments{
\item{inputTable}{data.frame containing the plate options}

\item{colGrowthCtl}{number of the column containing the growth control. Default is 11}

\item{colBrothCtl}{number of the column containing the broth control. Default is 12}

\item{nlines}{number of lines to read. Default is 8}
}
\description{
This function allows you to update an existing plate plan using the user's input
Useful to recalculate concentrations especially
}
\keyword{Plate}
\keyword{plan}
