% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineMICColumn.R
\name{determineMICcolumn}
\alias{determineMICcolumn}
\title{Convert growth/no growth tables to MIC column tables}
\usage{
determineMICcolumn(namedBinaryList, colGrowthCtl = 11,
  colBrothCtl = 12)
}
\arguments{
\item{namedBinaryList}{named list element from the output of convertGrowthToBinary()}

\item{colGrowthCtl}{number of the column containing the growth control. Default is 11}

\item{colBrothCtl}{number of the column containing the broth control. Default is 12}
}
\description{
This function allows you to convert a list of tables containing containing growth/no growth info from one
named list element to a list of tables containing the index of columns conatining the MIC concentration.
The output is a named list
}
\keyword{OD,}
\keyword{Tecan}
\keyword{growth}
\keyword{output,}
\keyword{percentage}
