% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucast_breakpoints.R
\name{eucast_breakpoints}
\alias{eucast_breakpoints}
\title{Determine EUCAST susceptibility}
\usage{
eucast_breakpoints(MIC.table, species.column = 1, MIC.column = 4,
  ATB.column = 3)
}
\arguments{
\item{MIC.table}{data.frame having MIC in one column.}

\item{species.column}{ID of the column containing the Species name recognized names are "A. baumannii", "P. aeruginosa", 
"E. coli" and "K. pneumoniae"}

\item{MIC.column}{ID of the column containing the MIC values in MIC.table. Defaults to 2}

\item{ATB.column}{ID of the column containing the ATB names in MIC.table. Defaults to 1}
}
\description{
This function given a data.frame with MIC will add a column to this table
giving the corresponding EUCAST susceptibility (Sensitive, Intermediate or Resistant)
}
\examples{
eucast_breakpoints(MIC.table)
}
\keyword{EUCAST}
\keyword{breakpoints}
\keyword{sensitivity,}
