% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractODTecanSheet.R
\name{extractODTecanSheet}
\alias{extractODTecanSheet}
\title{Extracts the OD values from Tecan Sheet data}
\usage{
extractODTecanSheet(tecanSheet, ODDataStartCol = 1, ODDataEndCol = 13)
}
\arguments{
\item{tecanSheet}{named list element from the output of splitSheetsTecanOutput()}

\item{ODDataStartCol}{number of the first column of the OD data. Default is 1}

\item{ODDataEndCol}{number of the last column of the OD data. Default is 13}
}
\description{
This function allows you to extract the OD values from one list element
containing the sheet data and the sheet name as produced by
splitSheetsTecanOutput()
It returns a list of data.frame() containing OD values read from each
excel sheet read with splitSheetsTecanOutput()
}
\keyword{OD,}
\keyword{Tecan}
\keyword{extraction}
\keyword{output,}
